refactor: Move quick_action_bar.rs to visualization module.
