+++ src/rust/ide/view/graph-editor/src/component/node.rs
@@ -5,7 +5,6 @@
 
 pub mod port;
 pub mod icon;
-pub mod quick_action_bar;
 
 pub use port::Expression;
