+++ src/rust/ide/view/graph-editor/src/component/visualization.rs
@@ -30,6 +30,7 @@ pub mod foreign;
 pub mod instance;
 pub mod path;
 pub mod registry;
+pub mod quick_action_bar;
 
 pub use container::*;
 pub use data::*;
